% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{char_map}
\alias{char_map}
\title{Split a Character Vector into its Unique Elements and a Mapping on These}
\usage{
char_map(x)
}
\arguments{
\item{x}{a character vector.}
}
\value{
a list with two elements \code{"chars"} and \code{"idx"} such that
  \code{ret$chars[ret$idx]} is identical to \code{x}.
}
\description{
Split a Character Vector into its Unique Elements and a Mapping on These
}
