% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcast.R
\name{dcast}
\alias{dcast}
\alias{agg}
\title{Data Frame Casting}
\usage{
dcast(X, by, ..., nthread = 2L)

agg(to, ..., to.keep = NULL, subset = NULL, to2char = paste,
  to2char.args = list(sep = "_"), names4res = paste,
  names4res.args = list(sep = "_"))
}
\arguments{
\item{X}{the data frame to cast.}

\item{by}{a character vector with one or more column names of \code{X} to
group by (the left hand side of the forumla in
\code{reshape2::\link[reshape2]{dcast}}/\code{data.table::\link[data.table]{dcast}}).}

\item{...}{for \code{dcast}, one or more objects as returned by \code{agg}.

  For \code{agg}, one or more unquoted expressions, each containing exactly
  one call to an aggregation function like \code{\bold{g}sum}. These
  expressions are evaluated in three stages:

  \bold{First}, the expression \code{x} passed as the first argument to the
  aggregation function is evaluated in the context of the data frame
  \code{X}. (See explanation on \code{x} below.) \bold{Second}, the
  aggregation function is called on all subsets (as defined by \code{by},
  \code{to}, \code{to.keep} and \code{subset}) of the vector resulting from
  the first stage. \bold{Third}, the complete expression is evaluated once
  for each corresponding \emph{result} column with the call to the
  aggregation function being replaced by the results of the aggregation
  function for this column.

  In the simplest case such a \code{...}-expression is just the call to the
  aggregation function, in which case there is no third stage.}

\item{nthread}{\emph{a hint} to the function on how many threads to use.}

\item{to}{a character vector with zero, one or more column names of \code{X}
by which to spread \code{X} (the right hand side of the forumla in
\code{reshape2::\link[reshape2]{dcast}}/\code{data.table::\link[data.table]{dcast}}).}

\item{to.keep}{a data.frame with one or more of the columns given as
\code{to}: keep only result columns corresponding to combinations of values
on the \code{to} columns that are contained in \code{to.keep}.}

\item{subset}{a quoted expression which is evaluated in the context of
\code{X} and must return a logical vector of length \code{nrow(X)}
indicating for each row of \code{X} whether it should be passed to the
aggregation function or not.}

\item{to2char}{vectorized function which is called with with the \code{k}
columns (each of length \code{n}) mentioned in \code{to} as the first
\code{k} arguments. It must return a character vector of length \code{n},
which is used as (part of) the column names of the result. This function is
only called if \code{length(to) > 0}.}

\item{to2char.args}{a list with additional arguments to pass to
\code{to2char}.}

\item{names4res}{a vectorized function which is called with the name of the
current \code{...}-argument of \code{agg} as the first argument and with
the result returned by \code{to2char} as the second argument. It must
return a character vector of the same length as its second argument, which
is used as the column names of the result. This function is only called if
the \code{...}-arguments are named and if \code{length(to) > 0}.}

\item{names4res.args}{a list with additional arguments to pass to
\code{names4res}.}
}
\description{
Data Frame Casting
}
