% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcast.R
\name{extract_agg_fun}
\alias{extract_agg_fun}
\title{Extract Details on the Call to an Aggregation Function from an Unevaluated
Expression}
\usage{
extract_agg_fun(x, agg_funs = names(fun.aggregates), pos = integer())
}
\arguments{
\item{x}{an unevaluate expression}

\item{agg_funs}{a character vector with the names of all aggregations
functions to look for}

\item{pos}{internal parameter, which must not be altered.}
}
\value{
If a call to an aggregation function was found inside the expression
  \code{x}, a list with the following components is returned: \code{agg_fun} -
  the name of the aggregation function found; \code{call} - the call to the
  aggregation function with all of the specified arguments being specified by
  their full names; \code{pos} - an integer giving the position of the call
  inside the expression \code{x}, i.e. \code{x[[pos]]} is the call to the
  aggregation function. If no call to an aggregation function was found,
  \code{NULL} is returned.
}
\description{
Extract Details on the Call to an Aggregation Function from an Unevaluated
Expression
}
\keyword{internal}
