% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gany}
\alias{gany}
\alias{gall}
\title{Aggregate Logical Vectors by Any/All}
\usage{
gany(x, na.rm = FALSE, fill = FALSE)

gall(x, na.rm = FALSE, fill = TRUE)
}
\arguments{
\item{x}{an expression to be evaluated in the context of the data frame
\code{X} to cast and resulting in an atomic vector of length
\code{nrow(X)}. In the simplest case this is
just an unqoted column name of the data frame \code{X}.}

\item{na.rm}{should \code{NA}s be removed from the vector resulting from the
evaluation of \code{x} before applying the aggregation
function to it?}

\item{fill}{value with which to fill empty cells in the result.}
}
\description{
Aggregate Logical Vectors by Any/All
}
