% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_fun.aggregate.R
\name{getSupport}
\alias{getSupport}
\title{Find the Best Fitting Support Declaration}
\usage{
getSupport(l, fun.aggregate, value.var)
}
\arguments{
\item{l}{the list of support declarations of the aggregation function}

\item{fun.aggregate}{the name of the aggregation function (just used for
error messages)}

\item{value.var}{the input vector to which to aggregation function should be
applied}
}
\value{
The best fitting support declaration given the \code{value.var}, i.e.
  an element of \code{l}.
}
\description{
Find the Best Fitting Support Declaration
}
\keyword{internal}
