% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_fun.aggregate.R
\name{register_fun.aggregate}
\alias{register_fun.aggregate}
\alias{support}
\title{Register an Aggregation Function with lfdcast}
\usage{
register_fun.aggregate(name, ptr, ..., rng = FALSE)

support(class, storage.mode, fill.default, fill.storage.modes,
  convert.fill.from, keep.attr)
}
\arguments{
\item{name}{the name of the aggregation function}

\item{ptr}{the external pointer to the function}

\item{...}{on or more objects as returned by \code{support}}

\item{class}{a character vector with the class(es) of inputs to which this
function can be applied. An \code{NA} elements means that it can be applied
to unclassed objects. A 0-length value for \code{class} means that all
classes are supported.}

\item{storage.mode}{a character vector with the storage mode(s) of inputs to
which this function can be applied}

\item{fill.default}{the default value with which to fill empty cells}

\item{fill.storage.modes}{a character vector with the storage mode(s)
supported for a user-supplied fill value}

\item{convert.fill.from}{a (potentially empty) character vector with storage
mode(s) (distinct from \code{fill.storage.modes}) which should be converted
to the storage mode \code{fill.storage.modes[1]}}

\item{keep.attr}{should the rsulting column(s) get the same attributes as the
input column? Can also be a character vector with the names of the
attributes to keep.}
}
\description{
Register an Aggregation Function with lfdcast
}
