% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{uniqueN}
\alias{uniqueN}
\title{Number of Unique Elements}
\usage{
uniqueN(x, na.rm = FALSE)
}
\arguments{
\item{x}{an atomic vector.}

\item{na.rm}{should \code{NA}s (and \code{NaN}s) be removed?}
}
\value{
the number of unique elements in \code{x}.
}
\description{
Number of Unique Elements
}
